+++ package.json
@@ -28,12 +28,12 @@
     "@nrwl/workspace": "15.6.3",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.7",
     "@solidjs/testing-library": "^0.7.0",
-    "@storybook/addon-essentials": "^7.5.0",
-    "@storybook/addon-interactions": "^7.0.10",
-    "@storybook/addon-links": "^7.0.10",
-    "@storybook/blocks": "^7.0.10",
-    "@storybook/html": "^7.0.10",
-    "@storybook/html-vite": "^7.5.0",
+    "@storybook/addon-essentials": "^7.6.10",
+    "@storybook/addon-interactions": "^7.6.10",
+    "@storybook/addon-links": "^7.6.10",
+    "@storybook/blocks": "^7.6.10",
+    "@storybook/html": "^7.6.10",
+    "@storybook/html-vite": "^7.6.10",
     "@storybook/testing-library": "^0.0.14-next.1",
     "@svgr/webpack": "^6.1.2",
     "@testing-library/jest-dom": "^5.16.5",
@@ -77,7 +77,7 @@
     "react-test-renderer": "^18.2.0",
     "rimraf": "^5.0.1",
     "s3-deploy": "^1.4.0",
-    "storybook": "^7.5.0",
+    "storybook": "^7.6.10",
     "tw-colors": "^1.2.6",
     "vite": "^4.1.4",
     "vite-plugin-dts": "^2.3.0",
